+++ .github/workflows/ccpp.yml
@@ -24,9 +24,9 @@ jobs:
           make -j$(nproc)
     - name: Bundle
       run: |
-        ./build/bundler -e ./build/bundler -e $(which chrpath) -e $(which strip) -d bundler
-        cp src/bundler bundler/bundler
-        tar -zcvf bundler_x86_64.tar.gz bundler
+        ./build/libtree -e ./build/libtree -e $(which chrpath) -e $(which strip) -d libtree
+        cp src/libtree libtree/libtree
+        tar -zcvf libtree_x86_64.tar.gz libtree
     - name: Create Release
       id: create_release
       uses: actions/create-release@v1
@@ -44,6 +44,6 @@ jobs:
         GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
       with:
         upload_url: ${{ steps.create_release.outputs.upload_url }}
-        asset_path: ./bundler_x86_64.tar.gz
-        asset_name: bundler_x86_64.tar.gz
+        asset_path: ./libtree_x86_64.tar.gz
+        asset_name: libtree_x86_64.tar.gz
         asset_content_type: application/gzip
