+++ README.md
@@ -1,17 +1,17 @@
-# bundler
+# libtree
 
 A tool that:
 - :deciduous_tree: turns `ldd` into a fancy tree
 - :point_up: explains why `ldd` finds shared libraries and why not
-- :package: optionally deploys relevant executables and dependencies into a single directory
+- :package: optionally deploys executables and dependencies into a single directory
 
 ## Installation
-Download the [**latest release**](https://github.com/haampie/bundler/releases) from Github:
+Download the [**latest release**](https://github.com/haampie/libtree/releases) from Github:
 
 ```
-- wget https://github.com/haampie/bundler/releases/download/v0.1.5/bundler_x86_64.tar.gz
-- tar -xzf bundler_x86_64.tar.gz
-- ./bundler/bundler -e $(which man)
+- wget https://github.com/haampie/libtree/releases/download/v1.0.0/libtree_x86_64.tar.gz
+- tar -xzf libtree_x86_64.tar.gz
+- ./libtree/libtree -e $(which man)
 ```
 
 ## Example 1: listing the dependencies of an executable
@@ -20,9 +20,9 @@ Download the [**latest release**](https://github.com/haampie/bundler/releases) f
 
 ## Example 2: deploying binaries + dependencies into a folder:
 ```bash
-$ bundler -e $(which bundler) -d bundler.bundle
+$ libtree -e $(which libtree) -d libtree.bundle
 Dependency tree
-bundler
+libtree
 ├── libcppglob.so.1 [runpath]
 │   ├── libstdc++.so.6 (skipped) [ld.so.conf]
 │   ├── libgcc_s.so.1 (skipped) [ld.so.conf]
@@ -31,16 +31,16 @@ bundler
 ├── libgcc_s.so.1 (skipped) [ld.so.conf]
 └── libc.so.6 (skipped) [ld.so.conf]
 
-Deploying to "bundler.bundle/usr"
-"/home/.../Documents/projects/bundler/build/bundler" => "bundler.bundle/usr/bin/bundler"
-"/home/.../Documents/projects/bundler/build/lib/libcppglob.so.1.1.0" => "bundler.bundle/usr/lib/libcppglob.so.1.1.0"
-  creating symlink "bundler.bundle/usr/lib/libcppglob.so.1"
+Deploying to "libtree.bundle/usr"
+"/home/.../Documents/projects/libtree/build/libtree" => "libtree.bundle/usr/bin/libtree"
+"/home/.../Documents/projects/libtree/build/lib/libcppglob.so.1.1.0" => "libtree.bundle/usr/lib/libcppglob.so.1.1.0"
+  creating symlink "libtree.bundle/usr/lib/libcppglob.so.1"
 
-$ tree bundler.bundle/
-bundler.bundle/
+$ tree libtree.bundle/
+libtree.bundle/
 └── usr
     ├── bin
-    │   └── bundler
+    │   └── libtree
     └── lib
         ├── libcppglob.so.1 -> libcppglob.so.1.1.0
         └── libcppglob.so.1.1.0
@@ -50,8 +50,8 @@ bundler.bundle/
 
 ## Build from source 
 ```bash
-git clone --recursive https://github.com/haampie/bundler.git
-cd bundler
+git clone --recursive https://github.com/haampie/libtree.git
+cd libtree
 mkdir build
 cd build
 cmake -DCMAKE_BUILD_TYPE=Release ..
