+++ include/libtree/deps.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <bundler/elf.hpp>
+#include <libtree/elf.hpp>
 
 #include <vector>
 #include <filesystem>
